Remove duplicate setup already present in core test helpers
`network/index` now exports all local mirage options to be used by
`boot` and the `setupApplication` core helper.
