+++ test/bigtest/network/index.js
@@ -1,9 +1,28 @@
-/* global require */
+import { camelize } from '@bigtest/mirage';
 
-let start = () => {}; // eslint-disable-line import/no-mutable-exports
+// auto-import all mirage submodules
+const req = require.context('./', true, /\.js$/);
+const modules = req.keys().reduce((acc, modulePath) => {
+  const moduleParts = modulePath.split('/');
+  const moduleType = moduleParts[1];
+  const moduleName = moduleParts[2];
 
-if (process.env.NODE_ENV !== 'production') {
-  start = require('./start').default; // eslint-disable-line global-require
-}
+  if (moduleName) {
+    const moduleKey = camelize(moduleName.replace('.js', ''));
 
-export default start;
+    return Object.assign(acc, {
+      [moduleType]: {
+        ...(acc[moduleType] || {}),
+        [moduleKey]: req(modulePath).default
+      }
+    });
+  } else if (modulePath === './config.js') {
+    return Object.assign(acc, {
+      baseConfig: req(modulePath).default
+    });
+  } else {
+    return acc;
+  }
+}, {});
+
+export default modules;
